move exercise 2 to loadImg
